move spotify client init to separate file
